Issue #860 SRP: separate empty topic list to new component and renames
