+++ src/components/topics/task_list_component.tsx
@@ -2,15 +2,13 @@
 import React from 'react';
 import { FlatList, ListRenderItemInfo } from 'react-native';
 import { History } from 'history';
-import { Trans } from '@lingui/react';
 import { TaskListItemComponent } from './task_list_item_component';
 import { Routes, goToRouteWithParameter } from '../../application/routing';
 import { Id, BookmarkTopicAction, UnbookmarkTopicAction } from '../../stores/topics';
-import { colors, values, getNormalFontFamily } from '../../application/styles';
+import { colors } from '../../application/styles';
 import { isTopicListHeading } from './is_topic_list_heading';
 import { ListItem } from './build_topic_list_items_with_headings';
 import { TopicListHeadingComponent } from './topic_list_heading_component';
-import { View, Text } from 'native-base';
 
 // tslint:disable-next-line:no-var-requires
 const R = require('ramda');
@@ -131,33 +129,4 @@ export class TaskListComponent extends React.PureComponent<Props, State> {
             />
         );
     }
-}
-
-export const NoTasksAddedComponent = (): JSX.Element => (
-    <EmptyTaskListComponent message={<Trans>No topics to show</Trans>} />
-);
-
-export const NoTasksRecommendedComponent = (): JSX.Element => (
-    <EmptyTaskListComponent message={<Trans>No topics to recommend</Trans>} />
-);
-
-export interface EmptyTaskListComponentProps {
-    readonly message: JSX.Element;
-}
-
-const EmptyTaskListComponent = (props: EmptyTaskListComponentProps): JSX.Element => (
-    <View
-        style={{
-            flex: 1,
-            paddingHorizontal: 10,
-            backgroundColor: colors.white,
-            borderRadius: values.lessRoundedBorderRadius,
-            margin: 5,
-            alignItems: 'center',
-        }}
-    >
-        <Text style={{ color: colors.darkerGrey, fontFamily: getNormalFontFamily() }}>
-            {props.message}
-        </Text>
-    </View>
-);
+};
